suppress warning: implicit conversion loses integer precision
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@44596 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ vm.c
@@ -2354,7 +2354,7 @@ nsdr(void)
 #include <execinfo.h>
 #define MAX_NATIVE_TRACE 1024
     static void *trace[MAX_NATIVE_TRACE];
-    int n = backtrace(trace, MAX_NATIVE_TRACE);
+    int n = (int)backtrace(trace, MAX_NATIVE_TRACE);
     char **syms = backtrace_symbols(trace, n);
     int i;
